feat: add KnownFields Node and pass into decode
After implementing a custom unmarshall, then using a Decode call on one of
nodes, the knownFields option is lost. This change allows passing it to the
decode call after the custom unmarshall.

There is a PR for this on the upstream, but it has not been merged the the
project is now in unmaintained state.

https://github.com/go-yaml/yaml/pull/1008